feat: [SOEN-52] Updated connection schema and routes

+++ src/server/api/routers/connections.ts
@@ -6,20 +6,45 @@ export const connectionsRouter = createTRPCRouter({
   getConnection: protectedProcedure
     .input(
       z.object({
-        user_1_ID: z.string(),
+        userEmail: z.string(),
       }),
     )
     .query(async ({ ctx, input }) => {
-      const user1ID = input.user_1_ID < ctx.session.user.id ? input.user_1_ID : ctx.session.user.id;
-      const user2ID = input.user_1_ID < ctx.session.user.id ? ctx.session.user.id : input.user_1_ID;
+      const recipient = await ctx.prisma.user.findUnique({
+        where: {
+          email: input.userEmail,
+        },
+      });
+      if (!recipient) {
+        throw new Error('User not found');
+      }
+
+      const user1ID = recipient.id < ctx.session.user.id ? recipient.id : ctx.session.user.id;
+      const user2ID = recipient.id < ctx.session.user.id ? ctx.session.user.id : recipient.id;
       const connection = await ctx.prisma.connection.findUnique({
         where: {
           user_1_ID_user_2_ID: { user_1_ID: user1ID, user_2_ID: user2ID },
         },
       });
       if (connection) {
+        if (
+          (ctx.session.user.id === user1ID && connection.connectionStatus === 'Pending_1_To_2') ||
+          (ctx.session.user.id === user2ID && connection.connectionStatus === 'Pending_2_To_1')
+        ) {
+          return {
+            status: 'Sent',
+          };
+        } else if (
+          (ctx.session.user.id === user1ID && connection.connectionStatus === 'Pending_2_To_1') ||
+          (ctx.session.user.id === user2ID && connection.connectionStatus === 'Pending_1_To_2')
+        ) {
+          return {
+            status: 'Received',
+          };
+        }
+
         return {
-          status: connection.connection_Status.includes('Pending') ? 'Pending' : connection.connection_Status,
+          status: connection.connectionStatus,
         };
       }
       return { status: '' };
@@ -27,38 +52,51 @@ export const connectionsRouter = createTRPCRouter({
   sendConnectionRequest: protectedProcedure
     .input(
       z.object({
-        user_1_ID: z.string(),
+        userEmail: z.string(),
       }),
     )
     .mutation(async ({ ctx, input }) => {
-      const user1ID = input.user_1_ID < ctx.session.user.id ? input.user_1_ID : ctx.session.user.id;
-      const user2ID = input.user_1_ID < ctx.session.user.id ? ctx.session.user.id : input.user_1_ID;
+      const recipient = await ctx.prisma.user.findUnique({
+        where: {
+          email: input.userEmail,
+        },
+      });
+      if (!recipient) {
+        throw new Error('User not found');
+      }
+
+      const user1ID = recipient.id < ctx.session.user.id ? recipient.id : ctx.session.user.id;
+      const user2ID = recipient.id < ctx.session.user.id ? ctx.session.user.id : recipient.id;
       return await ctx.prisma.connection.create({
         data: {
           user_1_ID: user1ID,
           user_2_ID: user2ID,
-          connection_Status: ctx.session.user.id == user1ID ? 'Pending_1_To_2' : 'Pending_2_To_1',
+          connectionStatus: ctx.session.user.id == user1ID ? 'Pending_1_To_2' : 'Pending_2_To_1',
         },
       });
     }),
 
-  addConnection: protectedProcedure
+  acceptConnection: protectedProcedure
     .input(
       z.object({
-        user_1_ID: z.string(),
+        userEmail: z.string(),
       }),
     )
     .mutation(async ({ ctx, input }) => {
-      const user1ID = input.user_1_ID < ctx.session.user.id ? input.user_1_ID : ctx.session.user.id;
-      const user2ID = input.user_1_ID < ctx.session.user.id ? ctx.session.user.id : input.user_1_ID;
-      return await ctx.prisma.connection.upsert({
-        create: {
-          user_1_ID: user1ID,
-          user_2_ID: user2ID,
-          connection_Status: 'Connected',
+      const recipient = await ctx.prisma.user.findUnique({
+        where: {
+          email: input.userEmail,
         },
-        update: {
-          connection_Status: 'Connected',
+      });
+      if (!recipient) {
+        throw new Error('User not found');
+      }
+
+      const user1ID = recipient.id < ctx.session.user.id ? recipient.id : ctx.session.user.id;
+      const user2ID = recipient.id < ctx.session.user.id ? ctx.session.user.id : recipient.id;
+      return await ctx.prisma.connection.update({
+        data: {
+          connectionStatus: 'Connected',
         },
         where: {
           user_1_ID_user_2_ID: { user_1_ID: user1ID, user_2_ID: user2ID },
@@ -69,12 +107,21 @@ export const connectionsRouter = createTRPCRouter({
   removeConnection: protectedProcedure
     .input(
       z.object({
-        user_1_ID: z.string(),
+        userEmail: z.string(),
       }),
     )
     .mutation(async ({ ctx, input }) => {
-      const user1ID = input.user_1_ID < ctx.session.user.id ? input.user_1_ID : ctx.session.user.id;
-      const user2ID = input.user_1_ID < ctx.session.user.id ? ctx.session.user.id : input.user_1_ID;
+      const recipient = await ctx.prisma.user.findUnique({
+        where: {
+          email: input.userEmail,
+        },
+      });
+      if (!recipient) {
+        throw new Error('User not found');
+      }
+
+      const user1ID = recipient.id < ctx.session.user.id ? recipient.id : ctx.session.user.id;
+      const user2ID = recipient.id < ctx.session.user.id ? ctx.session.user.id : recipient.id;
       return await ctx.prisma.connection.delete({
         where: {
           user_1_ID_user_2_ID: { user_1_ID: user1ID, user_2_ID: user2ID },
